requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.3
pyotp==2.9.0
colorama==0.4.6
sqlalchemy==2.0.41
aiosqlite==0.21.0
